Fixing signed-ness of xstd::zstd::status.

+++ includes/xstd/zstd.hpp
@@ -40,7 +40,7 @@ namespace xstd::zstd
 		FORCE_INLINE constexpr operator const size_t&() const noexcept { return value; }
 		FORCE_INLINE constexpr bool is_success() const noexcept { return int64_t( size_t( value ) ) >= 0; }
 		FORCE_INLINE constexpr bool is_error() const noexcept { return int64_t( size_t( value ) ) < 0; }
-		std::string to_string() const { return value >= 0 ? XSTD_ESTR( "Ok" ) : xstd::fmt::str( XSTD_ESTR( "ZSTD Error: -%llu" ), -value ); }
+		std::string to_string() const { return int64_t( value ) >= 0 ? XSTD_ESTR( "Ok" ) : xstd::fmt::str( XSTD_ESTR( "ZSTD Error: -%llu" ), uint64_t( -int64_t( value ) ) ); }
 	};
 
 	// Result type.
